chore: move ExecutionDrawer
