--- src/components/organisms/EntityDetails/ExecutionDrawer/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ExecutionDrawer';
