+++ src/components/organisms/index.ts
@@ -3,3 +3,4 @@ export {default as EntityListContent} from './EntityList/EntityListContent';
 export {default as PageBlueprint, PageHeader, PageWrapper, PageToolbar} from './PageBlueprint';
 export {ConditionFormItems} from './TriggersFormItems';
 export {ActionFormItems} from './TriggersFormItems';
+export {default as ExecutionDrawer, ExecutionDrawerHeader} from './ExecutionDrawer';
