+++ src/components/pages/TestSuites/TestSuiteDetails/TestSuiteExecution/TestSuiteExecutionDrawer.tsx
@@ -1,7 +1,7 @@
 import {FC} from 'react';
 
-import ExecutionDrawer from '@organisms/EntityDetails/ExecutionDrawer';
-import ExecutionDrawerHeader from '@organisms/EntityDetails/ExecutionDrawer/ExecutionDrawerHeader';
+import ExecutionDrawer from '@organisms/ExecutionDrawer';
+import ExecutionDrawerHeader from '@organisms/ExecutionDrawer/ExecutionDrawerHeader';
 
 import {useExecutionDetailsPick} from '@store/executionDetails';
