+++ src/components/pages/Tests/TestDetails/TestExecution/TestExecutionDrawer.tsx
@@ -1,7 +1,6 @@
 import {FC} from 'react';
 
-import ExecutionDrawer from '@organisms/EntityDetails/ExecutionDrawer';
-import ExecutionDrawerHeader from '@organisms/EntityDetails/ExecutionDrawer/ExecutionDrawerHeader';
+import {ExecutionDrawer, ExecutionDrawerHeader} from '@organisms';
 
 import {useExecutionDetailsPick} from '@store/executionDetails';
